requirements.txt
adal==0.5.0
asgiref==3.2.10
asn1crypto==1.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.11.4
chardet==3.0.4
cryptography==3.0
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
filelock==3.0.12
gcloud==0.18.3
google-api-core==0.1.4
google-api-python-client==1.6.5
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==0.28.0
google-cloud-speech==0.31.0
google-resumable-media==0.3.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
keyring==21.2.1
keystoneauth1==3.4.0
kiwisolver==1.2.0
matplotlib==2.2.2
msrest==0.4.26
msrestazure==0.4.21
numpy==1.19.0
oauth2client==4.1.2
oauthlib==2.0.6
pbr==5.4.5
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
rackspaceauth==0.8.1
requests==2.24.0
requests-oauthlib==0.8.0
rsa==4.6
six==1.15.0
social-auth-app-django==2.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==3.2.0
uritemplate==3.0.1
urllib3==1.25.9
webapp2==2.5.2
whitenoise==5.1.0
zipp==3.1.0
